Extract ConnectionDialog inline validation into a mixin
Move the 16-method inline form-field validation cluster verbatim out of the
ConnectionDialog god-object into a new ConnectionDialogValidationMixin
(sshpilot/connection_dialog_validation.py). ConnectionDialog inherits the
mixin, so self stays identical and the move is a pure cut-and-paste with no
logic change. The methods validate user-entered rows (nickname, host, port,
username) against self.validator and style them inline; none touch the SSH
connection/auth path.

Bases become (Adw.Window, ConnectionDialogValidationMixin) — the GObject base
stays first. Drop the now-unused ipaddress import from connection_dialog.py
(re is still used elsewhere). connection_dialog.py: 4515 -> 4211 lines.

Adds tests/test_connection_dialog_validation.py: extraction-contract checks
(ConnectionDialog inherits the mixin; each moved method is owned by it) plus
direct coverage of the port/host/nickname/field validation rules via a
lightweight fake row — these had only indirect coverage before.
